pyproject.toml
[tool.poetry]
name = "pet-project"
version = "0.1.0"
description = ""
authors = ["IvanStored <jus1stored@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
pydantic-settings = "^2.1.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^12.1.2"}
python-dotenv = "^1.0.0"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
loguru = "^0.7.2"
jinja2 = "^3.1.2"
requests = "^2.31.0"
bs4 = "^0.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
